Moved rss feeds tests
