+++ Tribler/Test/Core/Modules/RestApi/test_my_channel_endpoints.py
@@ -229,127 +229,6 @@ class TestMyChannelEndpoints(AbstractTestMyChannelEndpoints):
 
         return self.do_request('mychannel/torrents', expected_code=200).addCallback(verify_torrents_json)
 
-    @deferred(timeout=10)
-    def test_rss_feeds_endpoint_no_my_channel(self):
-        """
-        Testing whether the API returns the right JSON data if no channel has been created when fetching rss feeds
-        """
-        self.session.lm.channel_manager = ChannelManager(self.session)
-        return self.do_request('mychannel/rssfeeds', expected_code=404)
-
-
-class TestMyChannelRssEndpoints(AbstractTestMyChannelEndpoints):
-
-    @deferred(timeout=10)
-    def test_rss_feeds_endpoint_with_channel(self):
-        """
-        Testing whether the API returns the right JSON data if a rss feeds from a channel are fetched
-        """
-        expected_json = {u'rssfeeds': [{u'url': u'http://test1.com/feed.xml'}, {u'url': u'http://test2.com/feed.xml'}]}
-        channel_name = "my channel"
-        self.create_fake_channel(channel_name, "this is a short description")
-        channel_obj = self.session.lm.channel_manager.get_channel(channel_name)
-        for rss_item in expected_json[u'rssfeeds']:
-            channel_obj.create_rss_feed(rss_item[u'url'])
-
-        return self.do_request('mychannel/rssfeeds', expected_code=200, expected_json=expected_json)
-
-    @deferred(timeout=10)
-    def test_add_rss_feed_no_my_channel(self):
-        """
-        Testing whether the API returns a 404 if no channel has been created when adding a rss feed
-        """
-        self.session.lm.channel_manager = ChannelManager(self.session)
-        return self.do_request('mychannel/rssfeeds/http%3A%2F%2Frssfeed.com%2Frss.xml',
-                               expected_code=404, request_type='PUT')
-
-    @deferred(timeout=10)
-    def test_add_rss_feed_conflict(self):
-        """
-        Testing whether the API returns error 409 if a channel the rss feed already exists
-        """
-        expected_json = {"error": "this rss feed already exists"}
-        my_channel_id = self.create_fake_channel("my channel", "this is a short description")
-        channel_obj = self.session.lm.channel_manager.get_my_channel(my_channel_id)
-        channel_obj.create_rss_feed("http://rssfeed.com/rss.xml")
-
-        return self.do_request('mychannel/rssfeeds/http%3A%2F%2Frssfeed.com%2Frss.xml', expected_code=409,
-                               expected_json=expected_json, request_type='PUT')
-
-    @deferred(timeout=10)
-    def test_add_rss_feed_with_channel(self):
-        """
-        Testing whether the API returns a 200 if a channel has been created and when adding a rss feed
-        """
-
-        def verify_rss_added(_):
-            channel_obj = self.session.lm.channel_manager.get_my_channel(my_channel_id)
-            self.assertEqual(channel_obj.get_rss_feed_url_list(), ["http://rssfeed.com/rss.xml"])
-
-        expected_json = {"added": True}
-        my_channel_id = self.create_fake_channel("my channel", "this is a short description")
-        return self.do_request('mychannel/rssfeeds/http%3A%2F%2Frssfeed.com%2Frss.xml', expected_code=200,
-                               expected_json=expected_json, request_type='PUT')\
-            .addCallback(verify_rss_added)
-
-    @deferred(timeout=10)
-    def test_remove_rss_feed_no_channel(self):
-        """
-        Testing whether the API returns a 404 if no channel has been created when adding a rss feed
-        """
-        self.session.lm.channel_manager = ChannelManager(self.session)
-        return self.do_request('mychannel/rssfeeds/http%3A%2F%2Frssfeed.com%2Frss.xml',
-                               expected_code=404, request_type='DELETE')
-
-    @deferred(timeout=10)
-    def test_remove_rss_feed_invalid_url(self):
-        """
-        Testing whether the API returns a 404 and error if the url parameter does not exist in the existing feeds
-        """
-        expected_json = {"error": "this url is not added to your RSS feeds"}
-        self.create_fake_channel("my channel", "this is a short description")
-        return self.do_request('mychannel/rssfeeds/http%3A%2F%2Frssfeed.com%2Frss.xml', expected_code=404,
-                               expected_json=expected_json, request_type='DELETE')
-
-    @deferred(timeout=10)
-    def test_remove_rss_feed_with_channel(self):
-        """
-        Testing whether the API returns a 200 if a channel has been created and when removing a rss feed
-        """
-        def verify_rss_removed(_):
-            channel_obj = self.session.lm.channel_manager.get_my_channel(my_channel_id)
-            self.assertEqual(channel_obj.get_rss_feed_url_list(), [])
-
-        expected_json = {"removed": True}
-        my_channel_id = self.create_fake_channel("my channel", "this is a short description")
-        channel_obj = self.session.lm.channel_manager.get_my_channel(my_channel_id)
-        channel_obj.create_rss_feed("http://rssfeed.com/rss.xml")
-
-        return self.do_request('mychannel/rssfeeds/http%3A%2F%2Frssfeed.com%2Frss.xml', expected_code=200,
-                               expected_json=expected_json, request_type='DELETE').addCallback(verify_rss_removed)
-
-    @deferred(timeout=10)
-    def test_recheck_rss_feeds_no_channel(self):
-        """
-        Testing whether the API returns a 404 if no channel has been created when rechecking rss feeds
-        """
-        self.session.lm.channel_manager = ChannelManager(self.session)
-        return self.do_request('mychannel/recheckfeeds', expected_code=404, request_type='POST')
-
-    @deferred(timeout=10)
-    def test_recheck_rss_feeds(self):
-        """
-        Testing whether the API returns a 200 if the rss feeds are rechecked in your channel
-        """
-        expected_json = {"rechecked": True}
-        my_channel_id = self.create_fake_channel("my channel", "this is a short description")
-        channel_obj = self.session.lm.channel_manager.get_my_channel(my_channel_id)
-        channel_obj._is_created = True
-        channel_obj.create_rss_feed("http://rssfeed.com/rss.xml")
-
-        return self.do_request('mychannel/recheckfeeds', expected_code=200,
-                               expected_json=expected_json, request_type='POST')
-
 
 class TestMyChannelPlaylistEndpoints(AbstractTestMyChannelEndpoints):
